pyproject.toml
[project]
name = "dummy-bank"
version = "0.0.1"
description = ""
readme = "README.md"
authors = [{ name = "Rumee Ahmed", email = "rumeeahmad@gmail.com" }]
requires-python = "==3.14.*"

dependencies = [
  "asyncpg==0.31.0",
  "email-validator==2.3.0",
  "fastapi==0.119.1",
  "pydantic==2.12.5",
  "pydantic-settings==2.12.0",
  "pyjwt==2.11.0",
  "structlog==25.5.0",
  "sqlalchemy[asyncio]==2.0.46",
  "uvicorn==0.40.0",
]

[dependency-groups]
dev = [
  "alembic==1.18.3",
  "bandit==1.9.3",
  "coverage==7.13.3",
  "freezegun==1.5.5",
  "pytest==9.0.2",
  "pytest-asyncio==1.3.0",
  "pytest-xdist==3.8.0",
  "pytest-httpx==0.36.0",
  "python-dotenv==1.2.1",
  "pytest-env==1.2.0",
  "pytest-cov>=7.0.0",
  "pytest-postgresql==8.0.0",
  "psycopg==3.3.2",
  "ruff==0.15.0",
  "ty==0.0.15",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = ["wip", "integration"]
testpaths = ["tests"]

[tool.ruff.lint]
extend-select = ["I", "F401"]

[tool.ty.rules]
all = "error"

[tool.ty.analysis]
respect-type-ignore-comments = false

[tool.ty.environment]
python-version = "3.14"
root = ["."]

[tool.ty.src]
include = ["api", "domain", "lib", "repository", "tests", "migrations"]
exclude = ["__pycache__", ".pytest_cache", ".ruff_cache", "scratch"]

[tool.ty.terminal]
error-on-warning = true

[tool.coverage.run]
omit = [
  ".*",
  "*/site-packages/*",
  "*/tests/*",
  "*/conftest.py",
  "__init__.py",
  "*/settings.py",
]

[tool.coverage.report]
fail_under = 100
exclude_also = ["if __name__ == .__main__.:"]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self\\.debug",
  "if TYPE_CHECKING",
  "\\.\\.\\.",
]

[build-system]
requires = ["uv_build==0.9.18"]
build-backend = "uv_build"

[tool.uv-build]
packages = ["api", "domain", "repository", "lib", "migrations"]
